requirements.txt
torch
torchvision
timm
numpy
pandas
Pillow
matplotlib
tensorboard
